Support < 5 or > 5 options
When you click Edit, five blank rows will be added on the screen.
On save, any blank rows are removed in the backend code.

When you Submit after editing, numVotes will stay the same
based on option string value.  If you rename an option, count is lost.
If you change a string value and then enter the same string value in a
different row, the numVotes is preserved.

Note if there are duplicates, numVotes is only saved for the first one,
but it is a bug that you can have duplicate option strings